requirements.txt
boto>=2.7
bunch
decorator
distribute
envoy
logutils
pyyaml
requests
stevedore
